Add a new section on the "programs needed".
We've seen this many times before - people can get
python, but don't know what to use to write it.
So I've added a section on good code editors to use.

They're specifically not IDEs.

This also allowed reorganising the last section of the
page to be more compact and easy to follow (less broken
up by the dialogue boxes)